requirements.txt
langchain-core>=0.1.0
langgraph>=0.1.0
google-generativeai>=0.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
rich>=13.5.0
tavily-python>=0.1.1